+++ OpenGLCore/src/OpenGLCore/Graphics/Shader.h
@@ -30,7 +30,7 @@ namespace OpenGLCore::Graphics
 		void SetInt3(const std::string& name, const glm::ivec3& value);
 		void SetInt4(const std::string& name, const glm::ivec4& value);
 
-		void SetFloat(const std::string& name, int value);
+		void SetFloat(const std::string& name, float value);
 		void SetFloat2(const std::string& name, const glm::vec2& value);
 		void SetFloat3(const std::string& name, const glm::vec3& value);
 		void SetFloat4(const std::string& name, const glm::vec4& value);
